+++ src/app/projects/project/graph/components/chart/chart.data.ts
@@ -1,0 +1,7 @@
+import { SquishyNodeData } from '../squishy-node.data';
+import { ChartDatasetConfig } from './chart-dataset.config';
+
+export interface ChartData extends SquishyNodeData {
+    animation: boolean
+    datasets: { [key: string]: ChartDatasetConfig }
+}