+++ src/app/projects/project/graph/components/node-components.type.ts
@@ -2,5 +2,6 @@ export enum NodeComponentsType {
     FileInput = 'FileInput',
     FileOutput = 'FileOutput',
     Script = 'Script',
-    TextInput = 'TextInput'
+    TextInput = 'TextInput',
+    Chart = 'Chart'
 }