+++ src/app/projects/project/graph/components/node-dynamic-input.event.ts
@@ -1,0 +1,6 @@
+import { Node } from 'rete';
+
+export interface NodeDynamicInputEvent {
+    id: string
+    node: Node
+}