+++ src/worker/execution/node-executor/chart/chart.execution-data.ts
@@ -1,0 +1,1 @@
+export type ChartExecutionData = void